requirements.txt
pipecat-ai[openai,silero,deepgram,azure]
fastapi~=0.111.1
uvicorn~=0.30.3
python-dotenv~=1.0.1
loguru~=0.7.2
requests~=2.32.3
selenium~=4.23.0
aiohttp~=3.9.5
groq~=0.9.0
starlette~=0.37.2